+++ modules/windows-vm/main.tf
@@ -1,13 +1,22 @@
+locals {
+  tags = {
+    Size     = var.vm_size
+    Location = var.location
+    PublicIP = "ON"
+    Image    = var.storage_image_reference_sku
+  }
+}
+
 resource "azurerm_public_ip" "public" {
   name                = var.public_ip_name
   resource_group_name = var.resource_group_name
-  location            = var.resource_group_location
+  location            = var.location
   allocation_method   = "Static"
 }
 
 resource "azurerm_network_interface" "public" {
   name                = var.network_interface_name
-  location            = var.resource_group_location
+  location            = var.location
   resource_group_name = var.resource_group_name
 
   ip_configuration {
@@ -25,7 +34,7 @@ resource "azurerm_network_interface_security_group_association" "public" {
 
 resource "azurerm_virtual_machine" "public" {
   name                  = var.vm_name
-  location              = var.resource_group_location
+  location              = var.location
   resource_group_name   = var.resource_group_name
   network_interface_ids = [azurerm_network_interface.public.id]
   vm_size               = var.vm_size
@@ -60,6 +69,8 @@ resource "azurerm_virtual_machine" "public" {
     admin_username = var.os_profile_admin_username
     admin_password = var.os_profile_admin_password
   }
+
+  tags = length(var.tags) == 0 ? local.tags : var.tags
 
   depends_on = [
     azurerm_network_interface_security_group_association.public