+++ variables.tf
@@ -15,19 +15,3 @@ variable "subscription_id" {
   description = "Azure subscription ID"
   default     = "1b08b9a2-ac6d-4b86-8a2f-8fef552c8371"
 }
-
-variable "os_profile_admin_password" {
-  type        = string
-  description = "Windows OS admin user password"
-  default     = "2HiVkwYAx0VKJoAC"
-}
-
-variable "tags" {
-  type        = map(string)
-  description = "Tags for all resources"
-  default = {
-    Environment  = "DEV"
-    Owner        = "Terraform"
-    Autoshutdown = "OFF"
-  }
-}